+++ code/src/F0.CodeAnalysis.CSharp.Benchmarking/Inspection/Inspector.cs
@@ -1,4 +1,5 @@
 using System.Collections.Immutable;
+using System.Diagnostics;
 using System.Globalization;
 using System.Text;
 using F0.CodeAnalysis.CSharp.Diagnostics;
@@ -80,40 +81,6 @@ internal static class Inspector
 		}
 	}
 
-	internal static void Source(int index, (string HintName, string Source) expected, GeneratedSourceResult actual)
-	{
-		string source = actual.SourceText.ToString();
-		if (!actual.HintName.Equals(expected.HintName, StringComparison.Ordinal))
-		{
-			BenchmarkInspectionException.Throw($"Expected and actual hint name of source #{index} differ: ", expected.HintName, actual.HintName);
-		}
-
-		if (!source.Equals(expected.Source, StringComparison.Ordinal))
-		{
-			string diff = Diff.GetDiff(expected.Source, source);
-			string message = $"Expected and actual source text of source #{index} differ: " + Environment.NewLine + diff;
-			BenchmarkInspectionException.Throw(message);
-		}
-	}
-
-	internal static void Generator(ISourceGenerator expected, ISourceGenerator actual)
-	{
-		if (expected != actual)
-		{
-			BenchmarkInspectionException.Throw("Unexpected Generator:", expected.GetType(), actual.GetType());
-		}
-	}
-
-	internal static void GeneratedSources(int expected, ImmutableArray<GeneratedSourceResult> actual)
-	{
-		if (expected != actual.Length)
-		{
-			IEnumerable<string> messages = actual.Select(static generatedSource => "   - " + generatedSource.HintName);
-			string generatedSources = String.Join(Environment.NewLine, messages);
-			BenchmarkInspectionException.Throw("Unexpected number of diagnostics:", expected, actual.Length, generatedSources);
-		}
-	}
-
 	internal static void Exception(Exception? exception)
 	{
 		if (exception is not null)
@@ -127,11 +94,13 @@ internal static class Inspector
 		StringBuilder message = new($"Unexpected {nameof(Diagnostic)} #{index}:");
 		_ = message.AppendLine();
 
+		Debug.Assert(actual.Descriptor.CustomTags is ImmutableArray<string>);
+
 		bool @throw = Diff.WriteDiff(message, nameof(Diagnostic.Id), expected.Id, actual.Id)
 			| Diff.WriteDiff(message, nameof(Diagnostic.Descriptor.Category), expected.Category, actual.Descriptor.Category)
 			| Diff.WriteDiff(message, nameof(Diagnostic.Descriptor.Title), expected.Title, actual.Descriptor.Title)
 			| Diff.WriteDiff(message, nameof(Diagnostic.Descriptor.Description), expected.Description, actual.Descriptor.Description)
-			| Diff.WriteDiff(message, nameof(AdhocDiagnostic.Message), expected.Message, actual.GetMessage())
+			| Diff.WriteDiff(message, nameof(AdhocDiagnostic.Message), expected.Message, actual.GetMessage(null))
 			| Diff.WriteDiff(message, nameof(Diagnostic.Descriptor.MessageFormat), expected.MessageFormat, actual.Descriptor.MessageFormat)
 			| Diff.WriteDiff(message, nameof(Diagnostic.DefaultSeverity), expected.DefaultSeverity, actual.DefaultSeverity)
 			| Diff.WriteDiff(message, nameof(Diagnostic.Severity), expected.Severity, actual.Severity)
@@ -141,8 +110,8 @@ internal static class Inspector
 			| Diff.WriteDiff(message, nameof(Diagnostic.Descriptor.HelpLinkUri), expected.HelpLink, actual.Descriptor.HelpLinkUri)
 			| Diff.WriteDiff(message, nameof(Diagnostic.Descriptor.IsEnabledByDefault), expected.IsEnabledByDefault, actual.Descriptor.IsEnabledByDefault)
 			| Diff.WriteSequenceDiff(message, nameof(Diagnostic.AdditionalLocations), expected.AdditionalLocations, actual.AdditionalLocations, static expected => expected.SourceSpan, static actual => actual.SourceSpan)
-			| Diff.WriteSequenceDiff(message, nameof(Diagnostic.Descriptor.CustomTags), expected.CustomTags, actual.Descriptor.CustomTags)
-			| Diff.WriteSequenceDiff(message, nameof(Diagnostic.Properties), expected.Properties, actual.Properties, static expected => (expected.Key, expected.Value), static actual => (actual.Key, actual.Value));
+			| Diff.WriteSequenceDiff(message, nameof(Diagnostic.Descriptor.CustomTags), expected.CustomTags, actual.Descriptor.CustomTags.ToImmutableArray())
+			| Diff.WriteOrderedSequenceDiff(message, nameof(Diagnostic.Properties), expected.Properties, actual.Properties, static expected => (expected.Key, expected.Value), static actual => (actual.Key, actual.Value));
 
 		if (@throw)
 		{