+++ code/src/F0.CodeAnalysis.CSharp.Benchmarking/Properties/AssemblyInfo.cs
@@ -1,5 +1,7 @@
 using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("F0.CodeAnalysis.CSharp.Benchmarking.Tests")]
+[assembly: InternalsVisibleTo("F0.CodeAnalysis.CSharp.Analyzer.Benchmarking")]
 [assembly: InternalsVisibleTo("F0.CodeAnalysis.CSharp.IncrementalGenerator.Benchmarking")]
 [assembly: InternalsVisibleTo("F0.CodeAnalysis.CSharp.SourceGenerator.Benchmarking")]
+
+[assembly: InternalsVisibleTo("F0.CodeAnalysis.CSharp.Benchmarking.Tests")]